+++ packages/solid/src/toast/toast.tsx
@@ -12,10 +12,11 @@ export const Toast = (props: ToastProps) => {
 
   const api = useToastItem(toastParams)
   const rootProps = mergeProps(() => api().rootProps, localProps)
+  const customToast = api().render()
 
   return (
     <ToastItemProvider value={api}>
-      <ark.div {...rootProps} />
+      <ark.div {...rootProps}>{customToast || props.children}</ark.div>
     </ToastItemProvider>
   )
 }